Show alerts in a new window
This causes alerts to be shown in a new window instead of in the current
view controller. This has a few advantages. First, it prevents touches
from being seen by the view controller that is being covered by the
alert. This prevents weird behavior when an alert is presented over a
UITableViewController. Also, the new window exists above the navigation
bar and the keyboard, so the user won't be able to do weird things there
without dismissing the alert.